requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.24
Pillow==10.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
